+++ src/visitors/get-lambda-config.js
@@ -2,52 +2,85 @@ let parse = require('@architect/parser')
 let fs = require('fs')
 let path = require('path')
 
-/**
- * @param {Object} a - {}
- * @param {Array} b - ['runtime', 'nodejs12.x']
- * @returns {Object} - {runtime: 'nodejs12.x}
- */
-function invert(a, b) {
-  let isLayerOrPolicyArray = Array.isArray(b) && (b[0] === 'layers' || b[0] === 'policies')
-  let isLayerOrPolicyObject = typeof b === 'object' && (Object.keys(b)[0] === 'layers' || Object.keys(b)[0] === 'policies')
-  if (isLayerOrPolicyArray) {
-    let nom = b.shift()
-    if (!a[nom])
-      a[nom] = b
-    else
-      a[nom] = a[nom].concat(b)
-    return a
+function updateProps (props, config) {
+  for (let setting of config) {
+    let name
+    let value
+    let isArrayProp
+
+    /**
+     * Example
+     * ---
+     * policies foobar
+     * ---
+     * produces arc.aws: [[ 'policies', 'foobar' ]]
+     */
+    if (Array.isArray(setting)) {
+      // Normalize singular to AWS equivalents
+      if (setting[0] === 'policy') setting[0] = 'policies'
+      if (setting[0] === 'layer') setting[0] = 'layers'
+      name = setting[0]
+      value = setting[1]
+      isArrayProp = name === 'policies' || name === 'layers'
+    }
+
+    /**
+     * Example:
+     * ---
+     * policies
+     *   foobar
+     * ---
+     * produces arc.aws: [{ policies: [ 'foobar' ] }]
+     */
+    else if (typeof setting === 'object') {
+      // Normalize singular to AWS equivalents
+      if (setting.policy) {
+        setting.policies = setting.policy
+        delete setting.policy
+      }
+      if (setting.layer) {
+        setting.layers = setting.layer
+        delete setting.layer
+      }
+      name = Object.keys(setting)[0]
+      value = setting[name]
+      isArrayProp = name === 'policies' || name === 'layers'
+    }
+    else continue // Technically invalid and should have been caught by parser
+
+    // Populate default props with config
+    if (props[name] && isArrayProp) {
+      // Value may be a single item or an array
+      if (!Array.isArray(value)) value = [ value ]
+      props[name] = props[name].concat(value).filter(p => p)
+    }
+    else if (props[name] && typeof value !== 'undefined') {
+      props[name] = value
+    }
   }
-  if (isLayerOrPolicyObject) {
-    let nom = Object.keys(b)[0]
-    if (!a[nom])
-      a[nom] = Object.keys(b[nom])
-    else
-      a[nom] = a[nom].concat(Object.keys(b[nom]))
-    return a
-  }
-  a[b[0]] = b[1]
-  return a
+
+  return props
 }
+
+
 
 module.exports = function getPropertyHelper(arc, pathToCode) {
 
-  // default props
+  // Default props
   let props = {
     timeout: 5,
     memory: 1152,
     runtime: 'nodejs12.x',
     state: 'n/a',
     concurrency: 'unthrottled',
-    fifo: true,
+    fifo: true, // Applies only to queues
     layers: [],
     policies: [],
   }
 
-  // .arc global override
+  // Global settings
   if (arc.aws) {
-    let globals = arc.aws.reduce(invert, {})
-    props = Object.assign({}, props, globals)
+    props = updateProps(props, arc.aws)
   }
 
   // .arc-config local override
@@ -57,9 +90,24 @@ module.exports = function getPropertyHelper(arc, pathToCode) {
     let raw = fs.readFileSync(arcFile).toString().trim()
     let config = parse(raw)
     if (config.aws) {
-      config = config.aws.reduce(invert, {})
+      props = updateProps(props, config.aws)
     }
-    props = Object.assign({}, props, config)
+  }
+
+  // Layer validation
+  if (props.layers.length > 5) throw Error('Lambda can only be configured with up to 5 layers')
+  // CloudFormation fails without a helpful error if any layers aren't in the same region as the app because CloudFormation
+  if (props.layers.length) {
+    for (let layer of props.layers) {
+      let layerRegion = layer.split(':')[3]
+      let getRegion = s => s[0] === 'region' && s[1]
+      let arcRegion = arc.aws && arc.aws.some(getRegion) && arc.aws.find(getRegion)[1]
+      let region = process.env.AWS_REGION || arcRegion
+      if (region && region !== layerRegion) {
+        let msg = `Lambda layers must be in the same region as app\nApp region: ${region}\nLayer ARN: ${layer}`
+        throw Error(msg)
+      }
+    }
   }
 
   return function getProp(name) {