+++ changelog.md
@@ -4,10 +4,20 @@
 
 ## [1.3.4] 2020-06-09
 
+### Added
+
+- Added layer region validation (instead of letting CloudFormation fail without a helpful error)
+
+
 ### Changed
 
 - Removed `mkdirp` in favor of Node.js >= 10.x `mkdir` recursive
 - Updated dependencies
+
+
+### Fixed
+
+- Fixed `@aws` configuration in root project manifest and `.arc-config`, especially pertaining to the use of `layer` or `layers`; fixes #852, ht @jessrosenfield!
 
 ---